Refresh rivers and river stations data (#43)
https://eaflood.atlassian.net/browse/FSR-685

Note 1: switched to a tab separated file for rivers as some rivers have a
comma in the name
Note 2: processed the files from the format provided to remove dos line
endings, remove redundant columns and reorder the columns. The next set
of files should be in the correct format and should not need processing.